refactor(fetch_response): change topic_count and partition_count types from int8_t to int64_t for better scalability refactor(fetch_response): optimize writeRecordBatches method to calculate total size before writing batches chore(fetch_response): remove unused raw_data_size variable from RecordBatch structure style(kafka_server): clean up console output for fetch request processing style(log_metadata_reader): remove unused batch_count variable from readAllBatches method

+++ src/include/fetch_response.hpp
@@ -18,8 +18,8 @@ public:
 
   FetchResponse &writeHeader(int32_t correlation_id);
   FetchResponse &writeResponseData(int32_t throttle_time_ms, int16_t error_code,
-                                   int32_t session_id, int8_t topic_count);
-  FetchResponse &writeTopicHeader(uint128_t topic_id, int8_t partition_count);
+                                   int32_t session_id, int64_t topic_count);
+  FetchResponse &writeTopicHeader(uint128_t topic_id, int64_t partition_count);
 
   // Partition data writing methods
   FetchResponse &writePartitionData(
